python-app/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
openai